adds issue, column-field, converted-resource download scripts using file downloader
